requirements.txt
python-telegram-bot==21.0.1
httpx==0.28.1
